+++ R/tree_numbering.R
@@ -34,7 +34,7 @@
 #' @name Neworder
 #' @export
 NeworderPruningwise <- function(nTip, nNode, parent, child, nEdge) {
-  .C("ape_neworder_pruningwise", as.integer(nTip), as.integer(nNode),
+  .C(`ape_neworder_pruningwise`, as.integer(nTip), as.integer(nNode),
      as.integer(parent), as.integer(child), as.integer(nEdge),
      integer(nEdge))[[6]]
 }
@@ -47,7 +47,7 @@ NeworderPruningwise <- function(nTip, nNode, parent, child, nEdge) {
 #' @keywords internal
 #' @export
 NeworderPhylo <- function(nTip, parent, child, nEdge, whichwise) {
-  .C("ape_neworder_phylo", as.integer(nTip), as.integer(parent),
+  .C(`ape_neworder_phylo`, as.integer(nTip), as.integer(parent),
      as.integer(child), as.integer(nEdge), integer(nEdge),
      as.integer(whichwise), NAOK = TRUE)[[5]]
 }