+++ NEWS.md
@@ -1,4 +1,4 @@
-# TreeTools 1.9.0.9000
+# TreeTools 1.9.0.9001
 
 ## Improvements
 
@@ -6,6 +6,8 @@
 
 - `PhyDatToMatrix()` can (and by default does) override levels to write
   ambiguous tokens in custom formats such as `{01}`.
+
+- Call C functions using symbols, not strings.
 
 
 # TreeTools 1.9.0 #